feat(revit): extracts collection creation logic out into SendCollectionManager
to cleanup the root object builder
