+++ Converters/Revit/Speckle.Converters.RevitShared.DependencyInjection/RevitConverterModule.cs
@@ -4,7 +4,8 @@ using Speckle.Converters.Common;
 using Speckle.Converters.Common.DependencyInjection;
 using Speckle.Converters.RevitShared.Helpers;
 using Speckle.Converters.RevitShared.Services;
-using Speckle.Converters.RevitShared.ToSpeckle;
+
+// using Speckle.Converters.RevitShared.ToSpeckle;
 
 namespace Speckle.Converters.RevitShared.DependencyInjection;
 
@@ -38,7 +39,7 @@ public class RevitConverterModule : ISpeckleModule
     builder.AddScoped<ParameterValueExtractor>();
     builder.AddScoped<ParameterValueSetter>();
     builder.AddScoped<DisplayValueExtractor>();
-    builder.AddScoped<HostedElementConversionToSpeckle>();
+    // builder.AddScoped<HostedElementConversionToSpeckle>();
     builder.AddScoped<ParameterObjectAssigner>();
     builder.AddScoped<ISlopeArrowExtractor, SlopeArrowExtractor>();
